pyproject.toml
[project]
name = "ecos-backend"
version = "0.1.0"
authors = [
    { name = "Vadim Remezov", email = "remezov@cs.vsu.ru" },
]
description = "Back-end application for ecos project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg==0.30.*",
    "fastapi==0.115.*",
    "hatch==1.13.*",
    "pydantic-settings==2.7.*",
    "pydantic[email]==2.10.*",
    "uvicorn==0.32.*",
    "python-keycloak==5.1.*",
    "python-multipart==0.0.*",
    "streaming-form-data==1.19.*",
    "boto3==1.36.*",
    "python-magic==0.4.*",
    "fastapi-mail==1.4.*",
    "jinja2==3.1.*",
    "geoalchemy2[shapely]==0.17.*",
]

[dependency-groups]
dev = [
    "alembic==1.13.*",
    "pytest==8.3.*",
    "pre-commit==4.0.*",
]
lint = [
    "ruff==0.7.*",
]

[project.scripts]
run_server = "ecos_backend.main:__name__"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ecos_backend"]
